+++ test/integration/iaptResultsPage.js
@@ -128,6 +128,10 @@ describe('IAPT results page', () => {
       it('has a back link to the search page', () => {
         iExpect.backLinkContent($, `${constants.siteRoot}${routes.search.path}?query=${gpName}`);
       });
+
+      it(`should have the page title - 'Find psychological therapies services - There are no services available for \'${gpName}\'- NHS.UK`, () => {
+        expect($('head title').text()).to.equal(`Find psychological therapies services - There are no services available for \'${gpName}\' - NHS.UK`);
+      });
     });
 
     describe('zero results', () => {