+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Hyperlink search again link on no results page
 - Add extra info to search page about use of GP
 - Update content to latest tested
+- Give each page a unique title
 
 0.5.0 / 2018-08-02
 ==================